Move header labels to test folder
